+++ src/diff_driver.c
@@ -9,7 +9,6 @@
 #include "git2/attr.h"
 
 #include "diff.h"
-#include "diff_patch.h"
 #include "diff_driver.h"
 #include "strmap.h"
 #include "map.h"
